+++ src/main/java/wooteco/subway/dto/request/LineRequest.java
@@ -1,11 +1,13 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.request;
 
 import wooteco.subway.domain.Line;
 
 public class LineRequest {
     private String name;
     private String color;
-    
+
+    private LineRequest() {
+    }
 
     public LineRequest(String name, String color) {
         this.name = name;
