+++ src/main/java/wooteco/subway/dto/request/StationRequest.java
@@ -1,10 +1,12 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.request;
 
 import wooteco.subway.domain.Station;
 
 public class StationRequest {
     private String name;
-    
+
+    private StationRequest() {
+    }
 
     public StationRequest(String name) {
         this.name = name;
